+++ src/NovelSpider.Config/NovelSpider/Config/Configs.cs
@@ -1,4 +1,4 @@
-﻿using System;
+using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Reflection;
@@ -72,10 +72,10 @@ public class Configs
 		HaveFunction = "ZhanQunPinyinDir中译英";
 		UserID = Guid.NewGuid();
 		AssemblyVersion = Assembly.GetExecutingAssembly().GetName().Version;
-		DisplayVersion = "10.18.2-net10-test";
+		DisplayVersion = "10.18.5-net10-test";
 		TaskNovelInfo = new Dictionary<string, object>();
 		IsDemo = false;
-		Build = new DateTime(2026, 7, 9);
+		Build = new DateTime(2026, 8, 14);
 	}
 }