+++ src/NovelSpider.Local.Jieqi/Properties/AssemblyInfo.cs
@@ -1,4 +1,4 @@
-﻿using System.Diagnostics;
+using System.Diagnostics;
 using System.Reflection;
 using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
@@ -16,8 +16,8 @@ using System.Runtime.Versioning;
 [assembly: AssemblyTrademark("")]
 [assembly: AssemblyProduct("NovelSpider.Jieqi")]
 [assembly: AssemblyTitle("NovelSpider.Jieqi")]
-[assembly: AssemblyFileVersion("10.18.2.0")]
-[assembly: AssemblyVersion("10.18.2.0")]
+[assembly: AssemblyFileVersion("10.18.5.0")]
+[assembly: AssemblyVersion("10.18.5.0")]