+++ src/NovelSpider.Local/Properties/AssemblyInfo.cs
@@ -1,4 +1,4 @@
-﻿using System.Diagnostics;
+using System.Diagnostics;
 using System.Reflection;
 using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
@@ -16,8 +16,8 @@ using System.Runtime.Versioning;
 [assembly: AssemblyTrademark("")]
 [assembly: ComVisible(false)]
 [assembly: Guid("9fea8b3a-847c-4bf1-9d32-e0f6a22dc831")]
-[assembly: AssemblyFileVersion("10.18.2.0")]
-[assembly: AssemblyVersion("10.18.2.0")]
+[assembly: AssemblyFileVersion("10.18.5.0")]
+[assembly: AssemblyVersion("10.18.5.0")]